+++ qualang_tools/units/units.py
@@ -11,7 +11,7 @@ from qm import Program
 from inspect import stack
 from warnings import warn
 from typing import Union
-from numpy import round, ndarray
+from numpy import round, ndarray, log10, sqrt
 
 
 class _nanosecond:
@@ -193,11 +193,11 @@ class unit:
         data: Union[float, ndarray],
         duration: Union[float, int],
         single_demod: bool = False,
-    ):
+    ) -> Union[float, ndarray]:
         """Converts the demodulated data to volts.
 
         :param data: demodulated data. Must be a python variable or array.
-        :param duration: demodulation duration in ns.
+        :param duration: demodulation duration in ns. **WARNING**: this must be the duration of one slice in the case of ```demod.sliced``` and ```demod.accumulated```.
         :param single_demod: Flag to add the additional factor of 2 needed for single demod.
         :return: the demodulated data in volts.
         """
@@ -206,10 +206,34 @@ class unit:
         else:
             return 4096 * data * self.V / duration
 
-    def raw2volts(self, data: Union[float, ndarray]):
+    def raw2volts(self, data: Union[float, ndarray]) -> Union[float, ndarray]:
         """Converts the raw data to volts.
 
         :param data: raw data. Must be a python variable or array.
         :return: the raw data in volts.
         """
         return data * self.V / 4096
+
+    @staticmethod
+    def volts2dBm(
+        Vp: Union[float, ndarray], Z: Union[float, int] = 50
+    ) -> Union[float, ndarray]:
+        """Converts the peak voltage (amplitude) from volts to dBm.
+
+        :param Vp: the peak voltage (amplitude) in volts. Must be a python variable or array.
+        :param Z: the impedance. Default is 50 ohms.
+        :return: the power in dBm.
+        """
+        return 10 * log10(((Vp / sqrt(2)) ** 2 * 1000) / Z)
+
+    @staticmethod
+    def dBm2volts(
+        P_dBm: Union[float, ndarray], Z: Union[float, int] = 50
+    ) -> Union[float, ndarray]:
+        """Converts the power from dBm to volts (peak voltage or amplitude).
+
+        :param P_dBm: the power in dBm. Must be a python variable or array.
+        :param Z: the impedance. Default is 50 ohms.
+        :return: the corresponding peak voltage (amplitude) in volts.
+        """
+        return sqrt(2 * Z / 1000) * 10 ** (P_dBm / 20)